third_party/WebKit/LayoutTests/external/wpt/orientation-event/horizontal-surface-manual-expected.txt

This is a testharness.js-based test.
FAIL X angleassert_approx_equals: expected a number but got a "undefined"
FAIL Y angleassert_approx_equals: expected a number but got a "undefined"
Harness: the test ran to completion.
